Refactoring server into separate file
